backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiolimiter>=1.2.1",
    "elevenlabs>=2.29.0",
    "fastapi[standard]>=0.128.0",
    "instructor>=1.14.1",
    "ipykernel>=7.1.0",
    "openai>=2.14.0",
    "opencv-python>=4.12.0.88",
    "openrouter>=0.1.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
